Set a more stable rotational inertia for test.
When we fixed the RelativeJoint2D so that it correctly uses rotation (originally fixed here: https://github.com/erincatto/Box2D/issues/487) we started seeing an issue in a test-case where the test would "explode".  This was simply caused by the test bodies having super-low rotational inertia. This wasn't a problem before because the joint didn't correctly use it.  With the fix it did.

This is the reported case: https://issuetracker.unity3d.com/issues/physics-2d-gameobject-having-attached-reltaivejoint-moves-away-from-the-connected-rigidbody